Fix Win64 size_t/int conversion after b3f2277ea42a.
TBR=dcarney@chromium.org

Review URL: https://codereview.chromium.org/646983005

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@24720 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/register-allocator.cc
@@ -533,7 +533,7 @@ BitVector* RegisterAllocator::ComputeLiveOut(const InstructionBlock* block) {
   for (auto succ : block->successors()) {
     // Add values live on entry to the successor. Note the successor's
     // live_in will not be computed yet for backwards edges.
-    BitVector* live_in = live_in_sets_[succ.ToSize()];
+    BitVector* live_in = live_in_sets_[static_cast<int>(succ.ToSize())];
     if (live_in != NULL) live_out->Union(*live_in);
 
     // All phi input operands corresponding to this successor edge are live
